fix list user by project

+++ src/controller/project.controller.js
@@ -4,6 +4,24 @@ const Task = require("../model/Task");
 const projectService = require("../service/ProjectService");
 const { validationResult } = require("express-validator");
 const authService = require("../service/AuthToken");
+exports.getUserByProject = async (req, res) => {
+    const id = req.params.id;
+    const projects = Project.findOne({ _id: id })
+        .populate({ path: "users", select: "username" })
+        .then((project) => {
+            res.json({
+                code: "200",
+                message: "Danh sách projects",
+                body: project,
+            });
+        })
+        .catch((err) => {
+            res.status(500).json({
+                code: "500",
+                message: "Server error " + err.message,
+            });
+        });
+};
 exports.count = async (req, res) => {
     try {
         const id = req.params.id;
@@ -29,22 +47,22 @@ exports.count = async (req, res) => {
 //[PUT] api/project/status/:id
 exports.updateStatus = async (req, res) => {
     try {
-        if (req.role === "admin") {
-            const id = req.params.id;
-            const project = await Project.findOne({ _id: id });
-            project.status = project.status === "done" ? "undone" : "done";
-            await project.save();
-            res.status(200).json({
-                code: "200",
-                message: "Update status thành công!",
-                body: project,
-            });
-        } else {
-            res.status(200).json({
-                code: "400",
-                message: "Bạn không có quyền update status project",
-            });
-        }
+        // if (req.role === "admin") {
+        const id = req.params.id;
+        const project = await Project.findOne({ _id: id });
+        project.status = project.status === "done" ? "undone" : "done";
+        await project.save();
+        res.status(200).json({
+            code: "200",
+            message: "Update status thành công!",
+            body: project,
+        });
+        // } else {
+        //     res.status(200).json({
+        //         code: "400",
+        //         message: "Bạn không có quyền update status project",
+        //     });
+        // }
     } catch (err) {
         return res
             .status(500)
@@ -55,25 +73,23 @@ exports.updateStatus = async (req, res) => {
 //[put] api/project/:id
 exports.update = async (req, res) => {
     try {
-        if (req.role === "admin") {
-            const id = req.params.id;
-            const project = await Project.findOne({ _id: id });
-            if (!project || typeof project === "undefined") {
-                return res
-                    .status(200)
-                    .json({ message: "Project không tồn tại" });
-            }
-            const projectUpdate = await projectService.update(req);
-            res.status(200).json({
-                code: "200",
-                message: "Update Project thành công!",
-                projectUpdate,
-            });
-        } else {
-            res.status(200).json({
-                message: "Bạn không có quyền sửa thông tin project",
-            });
+        // if (req.role === "admin") {
+        const id = req.params.id;
+        const project = await Project.findOne({ _id: id });
+        if (project.length === 0 || typeof project === "undefined") {
+            return res.status(200).json({ message: "Project không tồn tại" });
         }
+        const projectUpdate = await projectService.update(req);
+        res.status(200).json({
+            code: "200",
+            message: "Update Project thành công!",
+            projectUpdate,
+        });
+        // } else {
+        //     res.status(200).json({
+        //         message: "Bạn không có quyền sửa thông tin project",
+        //     });
+        // }
     } catch (err) {
         return res.status(500).json("Server error " + err.message);
     }
@@ -81,25 +97,25 @@ exports.update = async (req, res) => {
 //[DELETE] /api/project/:id
 exports.delete = async (req, res) => {
     try {
-        if (req.role === "admin") {
-            const id = req.params.id;
-            const project = await Project.findOne({ _id: id });
-            if (!project) {
-                return res
-                    .status(200)
-                    .json({ code: "400", message: "Project không tồn tại" });
-            }
-            await Project.deleteOne({ _id: id });
-            res.status(200).json({
-                code: "200",
-                message: "Xóa project thành công!!",
-            });
-        } else {
-            res.status(200).json({
-                code: "400",
-                message: "Bạn không có quyền xóa project",
-            });
+        // if (req.role === "admin") {
+        const id = req.params.id;
+        const project = await Project.findOne({ _id: id });
+        if (project === 0 || typeof project === "undefined") {
+            return res
+                .status(200)
+                .json({ code: "400", message: "Project không tồn tại" });
         }
+        await Project.deleteOne({ _id: id });
+        res.status(200).json({
+            code: "200",
+            message: "Xóa project thành công!!",
+        });
+        // } else {
+        //     res.status(200).json({
+        //         code: "400",
+        //         message: "Bạn không có quyền xóa project",
+        //     });
+        // }
     } catch (err) {
         return res
             .status(500)
@@ -109,30 +125,30 @@ exports.delete = async (req, res) => {
 //[POST] /api/project
 exports.create = async (req, res) => {
     try {
-        if (req.role === "admin") {
-            const errors = validationResult(req);
-            if (!errors.isEmpty()) {
-                res.status(200).json({ errors: errors.array() });
-                return;
-            }
-            const project = await projectService.createService(req);
-            if (project === null) {
-                res.status(200).json({
-                    code: "400",
-                    message: "Thêm project không thành công",
-                });
-            }
-            res.status(200).json({
-                code: "200",
-                message: "Thêm project thành công",
-                project: project,
-            });
-        } else {
+        // if (req.role === "admin") {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            res.status(200).json({ errors: errors.array() });
+            return;
+        }
+        const project = await projectService.createService(req);
+        if (project === 0 || typeof project === "undefined") {
             res.status(200).json({
                 code: "400",
-                message: "Bạn không có quyền thêm project",
+                message: "Thêm project không thành công",
             });
         }
+        res.status(200).json({
+            code: "200",
+            message: "Thêm project thành công",
+            project: project,
+        });
+        // } else {
+        //     res.status(200).json({
+        //         code: "400",
+        //         message: "Bạn không có quyền thêm project",
+        //     });
+        // }
     } catch (err) {
         // log.error(`Create project error: ${err.message}`);
         return res.status(500).json("Server error " + err.message);
@@ -147,7 +163,7 @@ exports.show = async (req, res) => {
             Project.countDocuments({}),
             projectService.showAll(req),
         ]);
-        if (projects === null || typeof projects === "undefined") {
+        if (projects === 0 || typeof projects === "undefined") {
             res.staus(200).json({
                 code: "400",
                 message: "Danh sách project trống!!",
@@ -171,10 +187,10 @@ exports.addUser = async (req, res) => {
         const idUser = req.params.idUser;
         const user = await User.findOne({ _id: idUser });
         const project = await Project.findOne({ _id: idProject });
-        if (!project || typeof project === "undefined") {
+        if (project === 0 || typeof project === "undefined") {
             throw Error("Not found project");
         }
-        if (!user || typeof user === "undefined") {
+        if (user === 0 || typeof user === "undefined") {
             throw Error("Not found user");
         }
         if (!project.users.includes(idUser)) {
